[S05P13C105-107][BE] refactor: Team Id Integer로 변경

See merge request s05-webmobile2-sub3/S05P13C105!11

+++ backend/src/main/java/com/web/curation/team/TeamDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TeamDao extends JpaRepository<Team,Long> {
+public interface TeamDao extends JpaRepository<Team,Integer> {
 
     //모든 팀 조회
     List<Team> findAll();
